+++ src/hooks/socket/useChatSocket.ts
@@ -2,20 +2,20 @@ import { useCallback, useEffect, useRef, useState } from 'react';
 import SockJS from 'sockjs-client';
 import { CompatClient, Stomp } from '@stomp/stompjs';
 import { StompSubscription } from '@stomp/stompjs/src/stomp-subscription';
-import { Chat, Notification } from '../typings';
-import { SERVER_STOMP_URL } from '../apis';
-import { useDispatch, useSelector } from 'react-redux';
-import { RootState } from '../redux/store';
-import { append } from '../redux/features/notificationSlice';
+import { Chat } from '../../typings';
+import { SERVER_STOMP_URL } from '../../apis';
 
-export function useNotificationSocket(userId: number | string | undefined, callback?: (body: any) => void) {
+export function useChatSocket(roomId: number | string, callback?: () => void) {
   const socketRef = useRef<WebSocket | null>(null);
   const stompClientRef = useRef<CompatClient | null>(null);
   const subscriptionRef = useRef<StompSubscription | null | undefined>(null);
   const accessToken = window.localStorage.getItem('accessToken');
-  const dispatch = useDispatch();
-  const { isLogin } = useSelector((state: RootState) => state.user);
-  const notifications = useSelector((state: RootState) => state.notification.notifications);
+  const [chats, setChats] = useState<Chat[]>([]);
+  const firstChatRef = useRef<Chat | undefined>(undefined);
+
+  useEffect(() => {
+    callback && callback();
+  }, [chats]);
 
   useEffect(() => {
     if (!SERVER_STOMP_URL) return;
@@ -28,10 +28,9 @@ export function useNotificationSocket(userId: number | string | undefined, callb
     if (!socketRef || !stompClientRef) return;
 
     stompClientRef.current.connect({}, (receipt: any) => {
-      subscriptionRef.current = stompClientRef?.current?.subscribe(`/room/user/${userId}`, (message) => {
+      subscriptionRef.current = stompClientRef?.current?.subscribe(`/room/${roomId}`, (message) => {
         const body = JSON.parse(message.body);
-        dispatch(append(body));
-        callback && callback(body);
+        setChats((prev) => [...prev, body.body]);
       });
     });
 
@@ -39,20 +38,29 @@ export function useNotificationSocket(userId: number | string | undefined, callb
       subscriptionRef.current?.unsubscribe();
       stompClientRef.current?.disconnect();
     };
-  }, [isLogin]);
+  }, []);
+
+  const setFirstChatRef = useCallback((chat: Chat) => {
+    firstChatRef.current = chat;
+  }, []);
+
+  const send = useCallback((message: string) => {
+    const body = {
+      message,
+    };
+    stompClientRef?.current?.publish({
+      destination: `/send/${roomId}`,
+      headers: {
+        accessToken: accessToken || '',
+      },
+      body: JSON.stringify(body),
+    });
+  }, []);
 
   const unsubscribe = useCallback(() => {
     subscriptionRef.current?.unsubscribe();
     stompClientRef.current?.disconnect();
   }, []);
 
-  return { unsubscribe, notifications };
+  return { chats, setChats, firstChatRef, send, unsubscribe, setFirstChatRef };
 }
-
-/**
- * content: "ee님이 소켓연결테스트 에 참가신청하셨습니다"
- * createdAt: "2022-08-26T17:37:30.844"
- * id: 8
- * isread: false
- * postId: 14
- */
